backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
# Pin scientific stack to versions that build reliably on slim images
numpy==1.26.4
pandas==2.2.2
scipy==1.12.0
scikit-learn==1.4.2
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.6
supabase==2.3.0
beautifulsoup4==4.12.3
lxml==5.1.0
reportlab==4.0.7
Pillow==10.2.0
joblib==1.3.2
tensorflow==2.15.0; sys_platform != "win32"
twilio==8.10.0
resend==0.8.0
openai==1.12.0
PyPDF2==3.0.1
anthropic>=0.39.0
qrcode[pil]>=7.4.2
razorpay>=1.4.1
